==> notebooks/0-read-raw-data.Rmd <==
---
title: "1. Read raw data from desktop and save for the project"
author: "Maciej Beręsewicz"
output: html_notebook
---

Packages


```{r}
library(tidyverse)
library(readxl)
library(xtable)
library(countrycode)
```

Apprehensions for 2019

```{r}
excel_file <- "~/git/zbiory/cudzoziemcy/Straz graniczna/niel pobyt 2018-2019-szczegoly.xlsx"
excel_sheets(excel_file) %>%
  set_names() %>%
  map_df(~ read_excel(path = excel_file, sheet = .x, skip = 5, 
                      col_names = c("appreh", "place", "woj", "citizen_iso", "citizen_name", 
                                                  "women_0_17", "women_18_59", "women_60",
                                                  "men_0_17", "men_18_64", "men_65", "total"),
                      col_types = c(rep("text", 5), rep("numeric", 7))) %>%
           tidyr::fill(appreh, place, woj) %>%
           filter(str_detect(appreh, "Suma|RAZEM|zatrzymanie", negate = T)) %>%
           group_by(appreh, place, citizen_iso, citizen_name) %>%
           summarise_at(vars(women_0_17:men_65), sum, na.rm = T), .id = "half") %>%
  mutate(year = str_extract(half, "\\d{4}"),
         half = str_extract(half, "I{1,2}")) %>%
  ungroup() %>%
  filter(year == 2019) -> sg
```

```{r}
sg %>%
  ungroup() %>%
  count(year, half, appreh, place, wt = women_18_59+women_60+men_18_64+men_65) %>%
  spread(place, n, fill = 0) %>%
  mutate(appreh = ifelse(appreh == "tak", "yes", "no")) %>%
  select(year, half, `same year` = appreh, `within country`= kraj, `airports` = `lotnicza zewnętrzna`,
         `Ukraine` = `Ukraina`, `Russia` = Rosja, `Belarus` = `Białoruś`) %>%
  xtable(caption = "sfsfsa", label = "tab-sg-basic", digits = 0) %>%
  print.xtable(include.rownames = F, caption = "afasf", caption.placement = "top", format.args = list(big.mark = ","))
```

Police data for 2019

```{r}
map_df(1:4, 
       ~ read_excel(path = "~/git/zbiory/cudzoziemcy/Policja/Osoba-cudzoziemcy-rok 2019.xls",  
                         sheet = .x, 
                         skip = 3, 
                    col_names = c("id","age","sex","pesel","country","year","quarter","address_type")) %>%
  tidyr::fill(id, age, sex, country, year, quarter),
  #mutate(pesel_tak_nie = ifelse(is.na(pesel_tak_nie), "NIE", pesel_tak_nie)),
  .id = 'sheet') %>%
  mutate(sheet = case_when(sheet == 1 ~ "process",
                           sheet == 2 ~ "criminal",
                           sheet == 3 ~ "searching",
                           sheet == 4  ~ "traffic"),
         age  = age - 1) %>%
  group_by(sheet, id) %>%
  tidyr::fill(pesel)  %>%
  ungroup() -> police

police %>%
  mutate(half = ifelse(quarter %in% 1:2, "I", "II")) %>%
  distinct(sheet, id, pesel, half) %>%
  mutate(pesel = replace_na(pesel, "NIE"), 
         pesel= factor(pesel, c("TAK", "NIE"), c("Yes", "No"), ordered = T)) %>%
  count(year = as.character(2019), half, pesel, type = sheet) %>%
  spread(type, n, fill = 0) %>%
  select(year:pesel, process, searching, traffic, criminal) %>%
  xtable(caption = "sfsfsa", label = "tab-sg-basic", digits = 0) %>%
  print.xtable(include.rownames = F, caption = "afasf", caption.placement = "top", format.args = list(big.mark = ","))
```

PESEL data for 2019

1 = brak adresu w Polsce 2 = aktualny pobyt czasowy 3 = aktualny pobyt stały 4 = wymeldowny donikąd 5 = nieaktualny pobyt czasowy 6 = kraj zamieszkania poza Polską)
Płeć: 1-kob, 2-mężczy
Wiek
- podział na ekonomiczne grupy wieku: 1 = - wiek przedprodukcyjny — do 17 lat
- wiek produkcyjny: 2 = - dla męczyzn — 18-64, 3 = - dla kobiet — 18-59,
- wiek poprodukcyjny 4 =- dla mężczyzn — 65 i więcej, 5 = - dla kobiet — 60 i więcej.

Stan cywilny
1 = kawaler, panna; 2 = żonaty, zamężna; 3 = wdowiec, wdowa; 4 = rozwiedziony, rozwiedziona; null = brak danych

Woj
Miejsce zameldowania lub ostatniego zameldowania (jeżeli jest podane w rejestrze PESEL).
Kody w nawiasach
null = brak danych
DOLNOŚLĄSKIE (02) KUJAWSKO-POMORSKIE (04) LUBELSKIE (06) LUBUSKIE (08) ŁÓDZKIE (10) MAŁOPOLSKIE (12) MAZOWIECKIE (14) OPOLSKIE (16) PODKARPACKIE (18) PODLASKIE (20) POMORSKIE (22) ŚLĄSKIE (24) ŚWIĘTOKRZYSKIE (26) WARMIŃSKO-MAZURSKIE (28) WIELKOPOLSKIE (30) ZACHODNIOPOMORSKIE (32)

BYŁE WARSZAWSKIE (01) BYŁE BIALSKOPODLASKI (03) BYŁE BIAŁOSTOCKIE (05) BYŁE BIELSKIE (07) BYŁE BYDGOSKIE (09) BYŁE CHEŁMSKIE (11) BYŁE CIECHANOWSKIE (13) BYŁE CZĘSTOCHOWSKIE (15) BYŁE ELBLĄSKIE (17) BYŁE GDAŃSKIE (19) BYŁE GORZOWSKIE (21) BYŁE JELENIOGÓRSKIE (23) BYŁE KALISKIE (25) BYŁE KATOWICKIE (27) BYŁE KIELECKIE (29) BYŁE KONIŃSKIE (31) BYŁE KOSZALIŃSKIE (33) BYŁE KRAKOWSKIE (35) BYŁE KROŚNIEŃSKIE (37) BYŁE LEGNICKIE (39) BYŁE LESZCZYŃSKIE (41) BYŁE LUBELSKIE (43) BYŁE ŁOMŻYŃSKIE (45) BYŁE ŁÓDZKIE (47) BYŁE NOWOSĄDECKIE (49) BYŁE OLSZTYŃSKIE (51) BYŁE OPOLSKIE (53) BYŁE OSTROŁĘCKIE (55) BYŁE PILSKIE (57) BYŁE PIOTRKOWSKIE (59) BYŁE PŁOCKIE (61) BYŁE POZNAŃSKIE (63) BYŁE PRZEMYSKIE (65) BYŁE RADOMSKIE (67) BYŁE RZESZOWSKIE (69) BYŁE SIEDLECKIE (71) BYŁE SIERADZKIE (73) BYŁE SKIERNIEWICKIE (75) BYŁE SŁUPSKIE (77) BYŁE SUWALSKIE (79) BYŁE SZCZECIŃSKIE (81) BYŁE TARNOBRZESKIE (83) BYŁE TARNOWSKIE (85) BYŁE TORUŃSKIE (87) BYŁE WAŁBRZYSKIE (89) BYŁE WŁOCŁAWSKIE (91) BYŁE WROCŁAWSKIE (93) BYŁE ZAMOJSKIE (95) BYŁE ZIELONOGÓRSKIE (97) NIEZNANE (99)


```{r}
map_df(1:4, ~ read_excel(path ="~/git/zbiory/cudzoziemcy/PESEL/cudzoziemcy-pesel-2019.xlsx",
                            sheet = .x, 
                            .name_repair = janitor::make_clean_names, range = cell_cols(1:14))) -> pesel
```

```{r}
pesel %>%
  count(as_of = stan_na, status_zameldowania, wt = liczba_cudzoziemcow) %>%
  mutate(status_zameldowania = factor(status_zameldowania, 1:6, 
                                      c("No address in Poland", "Temporal", "Permanent", 
                                        "Deregistered", "Out-of-date temporal", "Outside Poland"))) %>%
  spread(status_zameldowania, n) %>%
  xtable(caption = "sfsfsa", label = "tab-sg-basic", digits = 0) %>%
  print.xtable(include.rownames = F, caption = "afasf", caption.placement = "top", format.args = list(big.mark = ","))
```


```{r}
pesel %>%
  filter(status_zameldowania %in% c(2:3)) %>%
  count(quarter = stan_na, country = nazwa_kraju_obywatelstwa, country_code = kod_kraju_obywatelstwa, 
        sex = plec, age = wiek, wt = liczba_cudzoziemcow) -> pesel_aggr
```


Joining the data

- align names
- align ages
- fix names
- join only non-EU citizens

```{r}
dict_full <- read_delim("../data-raw/country-dictionary.csv", delim = ";") %>%
  mutate(iso3n = str_pad(iso3n, width = 3, pad = "0",side = "left"))
```


```{r}
sg %>%
  select(-women_0_17,-men_0_17) %>%
  gather(sex, border, women_18_59:men_65) %>%
  separate(sex, into = c("sex", "age"), sep = "_", extra = "merge") %>%
  filter(border > 0, appreh == "nie", place == "kraj") %>%
  select(year, half, country_name = citizen_name, country_iso = citizen_iso, sex, age, border) %>%
  mutate(country_iso3n = countrycode(country_iso, "iso3c", "iso3n"),
         age = ifelse(age %in% c("60", "65"), 2, 1),
         sex = ifelse(sex == "women", 1, 2),
         country_name = tolower(country_name),
         country_iso3n = str_pad(country_iso3n, 3, "left", 0),
         country_iso3n = ifelse(country_name == "kosovo", "567", country_iso3n),
         country_iso3n = replace_na(country_iso3n, "999")) %>%
  ungroup() %>%
  select(half,sex, age, country_name, country_iso3n, country_iso, border) -> sg_prepared
```

```{r}
saveRDS(object = sg_prepared, file = "../data/data-border-guards.rds")
```

```{r}
police %>%
  filter(age >= 18, sex != "Nieznana") %>%
  mutate(sex = ifelse(sex == "Mężczyzna", 2, 1),
         age = case_when(sex == 1 & age >= 60 ~ 2,
                         sex == 2 & age >= 65 ~ 2,
                         TRUE ~ 1),
         half = ifelse(quarter %in% 1:2, "I", "II"),
         registered = !is.na(pesel),
         country_name = tolower(country)) %>%
  count(id, sex, age, registered, country_name, half) %>%
  select(-n) %>%
  count(sex, age, registered, country_name, half) %>%
  spread(registered, n, fill = 0) %>%
  rename(police_nonreg = `FALSE`, police_reg = `TRUE`) %>%
  mutate(police_all = police_nonreg + police_reg) %>%
  left_join(dict_full %>% select(country_name=country, country_iso3n=iso3n)) %>%
  mutate(country_iso3n = case_when(
    country_name  == "aruba" ~ "533",
    country_name  %in% c("czechosłowacja", "jugosławia") ~ "999",
    country_name  == "jugosławia (serbia i czarnogóra)" ~ "688", 
    country_name  == "gwinea bissau" ~ "624",
    country_name  == "hongkong" ~ "344",
    country_name  == "kostaryka" ~ "188",
    country_name  == "laos" ~ "418",
    country_name  == "madagaskar" ~ "450",
    country_name  == "malezja" ~ "458",
    country_name  == "mauretania" ~ "478",
    country_name == "namibia" ~ "516",
    country_name == "panama" ~ "591",
    country_name == "ruanda" ~ "646",
    country_name == "salwador" ~ "222",
    country_name == "tajlandia (syjam)" ~ "764",
    country_name == "wyspy świętego tomasza i książęca" ~ "678",
    country_name == "wyspy zielonego przylądka" ~ "132",
    country_name == "andora" ~ "020",
    country_name == "antigua i barbuda" ~ "028",
    country_name == "kambodża (kampucza)" ~ "116",
    TRUE ~ country_iso3n
  ))   -> police_prepared
```

```{r}
saveRDS(object = police_prepared, file = "../data/data-police.rds")
```


```{r}
# different codding
pesel_aggr %>%
  filter(age != 1) %>%
  mutate(half = case_when(quarter == "2019.03.31" ~ "I",
                          quarter == "2019.06.30" ~ "Ia",
                          quarter == "2019.09.30" ~ "II",
                          quarter == "2019.12.31" ~ "IIa"),
         country_name = tolower(country),
         age = ifelse(age %in% 2:3, 1, 2)) %>%
  count(half, country_name, age, sex, wt = n, name = "pesel") %>%
  left_join(dict_full %>% filter(source3 == "pesel") %>% select(country_name=country, country_iso3n=iso3n)) %>%
  mutate(country_iso3n = case_when(country_name == "malediwy" ~ "462",
                                   country_name == "saint vincent i grenadyny" ~ "670",
                                   country_name == "suazi" ~ "748",
                                   country_name == "vanuatu" ~ "548",
                                   TRUE ~ country_iso3n )) -> pesel_prepared

pesel_prepared
```

```{r}
saveRDS(object = pesel_prepared, file = "../data/data-pesel.rds")
```

```{r}
pesel_prepared %>%
  filter(half %in% c("I", "II")) %>%
  mutate(country_iso3n = ifelse(country_iso3n == "998", "999", country_iso3n)) %>%
  count(half, country_iso3n, age, sex, wt = pesel, name = "pesel") %>%
  full_join(
    sg_prepared %>% 
      mutate(country_iso3n = ifelse(country_iso3n == "998", "999", country_iso3n)) %>%
      count(half, country_iso3n, age, sex, wt = border, name = "border")
    ) %>%
  full_join(
    police_prepared %>% 
      mutate(country_iso3n = ifelse(country_iso3n == "998", "999", country_iso3n)) %>%
      group_by(half,  country_iso3n, age, sex) %>%
      summarise_at(vars(police_nonreg:police_all), sum) %>%
      ungroup()
    )  %>%
  select(half, age, sex, country_iso3n, pesel, border, matches("police")) %>%
  mutate(eu = countrycode(country_iso3n, "genc3n", "eu28"),
         country_short = countrycode(country_iso3n, "genc3n", "iso3c"),
         country_continent = countrycode(country_iso3n, "genc3n", "continent"),
         country = countrycode(country_iso3n, "genc3n", "un.name.en"),
         country_efta = country %in% c("Iceland", "Norway", "Liechtenstein", "Switzerland"),
         country = case_when(country_iso3n == "999" ~ "other",
                             country_iso3n == "998" ~ "other",
                             country_iso3n == "567" ~ "Kosovo",
                             country_iso3n == "275" ~ "Palestine",
                             country_iso3n == "158" ~ "Taiwan",
                             country_iso3n == "336" ~ "Vatican city",
                             country_iso3n == "016" ~ "American Samoa",
                             TRUE ~ country),
         country_continent = case_when(country == "Kosovo" ~ "Europe",
                                       country == "Palestine" ~ "Asia",
                                       country == "other" ~ "Rest",
                                       TRUE ~ country_continent),
         country_short = ifelse(is.na(country_short), "UNK", country_short)) %>%
  mutate_at(vars(pesel:police_all), replace_na, 0)  %>%
  filter(!is.na(country)) %>%
  arrange(country, country_iso3n, country_short,country_continent, country_efta, eu, 
          half, age, sex, pesel, border, police_nonreg, police_reg, police_all) %>%
  select(country, country_iso3n, country_short, country_continent, country_efta, eu, 
         half, age, sex, pesel, border, police_nonreg, police_reg, police_all)  -> countries_all

countries_all  
```

```{r}
saveRDS(countries_all, file = "../data/data-for-model.rds")
```
