==> notebooks/1-zhang-model.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
library(maxLik)
library(sandwich)
library(lmtest)
```

```{r}
lichun_ll <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  phi <- param[3]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  mu <- N^alpha*(n/N)^beta
  res <- m * log(mu) - (m + phi) * log(mu + phi) + (m + phi - 0.5) * log(m + phi) + 0.5 * log(phi)
  res
}

lichun_ll_grad <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  phi <- param[3]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  mu <- N^alpha*(n/N)^beta
  x <- cbind(log(N),log(n/N))
  d_gamma <- as.numeric((m - mu) / (mu + phi)) * (phi* x )
  d_phi <- -log(mu + phi) - (m + phi)/(mu + phi) + log(m + phi) + (m + phi -0.5) / (m + phi) + 1/(2*phi)
  cbind(d_gamma, d_phi)
  #c(alpha=d_gamma[1], beta = d_gamma[2], phi=sum(d_phi))
}


lichun_ll_hess <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  phi <- param[3]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  mu <- N^alpha*(n/N)^beta
  x <- cbind(log(N),log(n/N))
  
  #  
  d2_gamma <-  t(x) %*% (x * -as.numeric( (m+phi)/(mu+phi)*phi ))
  d2_phi <- sum( -(2*mu+phi-m)/(mu+phi)^2 + (m+phi +0.5)/(m+phi)^2 - 1/(2*phi^2) )
  d2_gamma_phi <- colSums( -as.numeric( (mu - m) / (mu+phi)^2 * mu) * x)
  
  h <- as.matrix(Matrix::bdiag(d2_gamma, d2_phi))
  colnames(h) <- rownames(h) <- c("alpha","beta","phi")
  h[3,1:2] <- h[1:2,3] <- d2_gamma_phi
  h
}
```


```{r}
dt2019_1 <- m %>% filter(half == "I") %>%
  rename(m = border,
         N = pesel,
         n = police_all)
```


```{r}
m0 <- lm(log(m/N) ~ -1 + log(N) + log(n/N), data = dt2019_1)
summary(m0)
plot(m0)
```

```{r}
wynik <- maxLik(logLik = ll, 
                grad = grad, 
                hess = hess,
                data = dt2019_1, 
                method = "NR",
                start = c(alpha = coef(m0)[1]+1, beta = coef(m0)[2], phi = 1/sigma(m0)))
summary(wynik)
AIC(wynik)
confint(wynik, level = 0.95)

sum(dt2019_1$N^0.6139188)
sum(dt2019_1$N^coef(wynik)[1])
sum(dt2019_1$N^0.7165990)
```

```{r}

approx_log_fac <- function(n){ 
    a <- n*log(n) - n + log(pi*(1+4*pi*(1+2*n)))/6 + log(pi)/2
    a
}

gammabrob <- function(x) as.brob(gamma(x))


```

```{r}
poisson_ll <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  mu <- N^alpha*(n/N)^beta
  res <- dpois(x = m, lambda = mu, log = T)
  #res <- -mu + m*log(mu)
  res
}

poisson_grad <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  n_N <- n/N
  res_alpha <- -log(N)*n_N^beta*N^alpha + m*log(N) 
  res_beta <- -log(n_N)*N^alpha*(n_N^beta) + m*log(n_N)
  res <- cbind(res_alpha, res_beta)
  res
}
```

```{r}
tpoisson_ll <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  mu <- N^alpha*(n/N)^beta
  res <- dtpois(x = m, lambda = mu, log = T, a = 0)
  #res <- -mu + m*log(mu)
  res
}

t2poisson_ll <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  mu <- N^alpha*(n/N)^beta
  res <- dtpois(x = m, lambda = mu, log = T, a = 1)
  #res <- -mu + m*log(mu)
  res
}

```

## Negative binomial

```{r}
negbin_ll <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  phi <- param[3]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  mu <- N^alpha*(n/N)^beta
  res <- dnbinom(x = m, size = phi, mu = mu, log = T)
  res
}
```

```{r}
negbin_ll <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  phi <- param[3]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  mu <- N^alpha*(n/N)^beta
  res <- lgamma(m+phi^-1) - lgamma(phi^-1) - approx_log_fac(m) - (m + phi^-1)*log(1+phi*mu) + m*log(phi*mu) 
  res
}
```

## Zero-truncated Negative Binomial


```{r}
ztnegbin_ll <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  phi <- param[3]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  mu <- N^alpha*(n/N)^beta
  s <- phi^-1
  res <- lgamma(m+s) - lgamma(s) - approx_log_fac(m) -  
    (m + s)*log(1+phi*mu) +  m*log(phi*mu) - log(1 - (1+phi*mu)^(-s))
  res
}

ztnegbin_ll_grad <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  phi <- param[3]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  mu <- as.numeric(N^alpha*(n/N)^beta)
  v <- 1 + phi*mu
  s <- phi^-1
  c <- length(m)
  d_phi <- numeric(c)
  x <- cbind(log(N),log(n/N))
  for (i in 1:c) {
    d_phi[i] <- sum(1/(phi*(0:(m[i]-1)) + 1)) + 
      m[i] + 
      log(v[i])/phi - 
      mu[i]*(phi*m[i]+1)/v[i] + 
      (v[i]*log(v[i])-phi*mu[i])/(phi*v[i]*(v[i]^s-1))
  }
  d_gamma <- -mu*x*as.numeric(1/(v*v^s-1) - (m+mu)/(mu*v))
  cbind(d_gamma, d_phi)
}

ztnegbin_ll_hess <- function(param, data) {
  alpha <- param[1]
  beta <- param[2]
  phi <- param[3]
  m <- as.matrix(data$m)
  n <- as.matrix(data$n)
  N <- as.matrix(data$N)
  mu <- as.numeric(N^alpha*(n/N)^beta)
  v <- 1 + phi*mu
  s <- phi^-1
  c <- length(m)
  d2_phi <- numeric(c)
  x <- cbind(log(N),log(n/N))

  for (i in 1:c) {
    d2_phi[i] <- sum(phi*(0:(m[i]-1))/(phi*(0:(m[i]-1)) + 1)^2) + 
      log(v[i])^2*v[i]^s/(phi^2*(v[i]^s-1)^2) - 
      log(v[i])*(2*mu[i]*v[i]^s+v[i]^s+phi*mu[i]*(v[i]^s-1)-1)/(phi*v[i]*(v[i]^s-1)^2) -
      log(v[i])/phi + 
      (phi*mu[i]*(2*mu[i]-m[i])+mu[i])/v[i]^2 + 
      (mu[i]*(mu[i]*v[i]^s+v[i]^s+2*phi*mu[i]*(v[i]^s-1)-1)) / (v[i]^2*(v[i]^s-1)^2)
  }
  
  d2_gamma <- t(x) %*% (x *(mu* as.numeric((v^s*(mu-1)+1)/(v^2*(v^s-1)^2) - (phi*m+1)/v^2)))
  d2_gamma_phi <- mu*x*as.numeric((mu*(v^s + phi*(v^s-1)) - s*v^(s+1)*log(v))/(v^2*(v^s-1)^2) + (phi*(mu-m))/v^2)
  
  h <- as.matrix(Matrix::bdiag(d2_gamma, sum(d2_phi)))
  colnames(h) <- rownames(h) <- c("alpha","beta","phi")
  h[3,1:2] <- h[1:2,3] <- colSums(d2_gamma_phi)
  h
  
}

```

```{r}
m_start <- lm(log(m/N) ~ -1 + log(N) + log(n/N), data = dt2019_1)
summary(m_start)
plot(m_start)
```

```{r}
wynik_poisson <- maxLik(logLik = poisson_ll, grad = poisson_grad,
                       data = dt2019_1, method = "BFGS",
                       start = c(alpha = unname(coef(m_start)[1] + 1),
                                 beta = unname(coef(m_start)[2])))

wynik_tpoisson <- maxLik(logLik = tpoisson_ll,
                       data = dt2019_1, method = "BFGS",
                       start = c(alpha = unname(coef(m_start)[1] + 1),
                                 beta = unname(coef(m_start)[2])))

wynik_t2poisson <- maxLik(logLik = t2poisson_ll,
                       data = subset(dt2019_1, m > 1), method = "BFGS",
                       start = c(alpha = unname(coef(m_start)[1] + 1),
                                 beta = unname(coef(m_start)[2])))


wynik_lichun <- maxLik(logLik = lichun_ll, grad = lichun_ll_grad, hess = lichun_ll_hess,
                      data = dt2019_1,
                      method = "BFGS",
                      start = c(alpha = unname(coef(m_start)[1] + 1) ,
                                beta = unname(coef(m_start)[2]),
                                phi = 1/sigma(m_start)))

wynik_negbin <- maxLik(logLik = negbin_ll,
                      data = dt2019_1, 
                      method = "BFGS",
                      start = c(alpha = unname(coef(m_start)[1] + 1) ,
                                beta = unname(coef(m_start)[2]), 
                                phi = 1/sigma(m_start)))

wynik_ztnegbin <- maxLik(logLik = ztnegbin_ll, 
                         grad = ztnegbin_ll_grad, hess = ztnegbin_ll_hess,
                      data = dt2019_1,
                      method = "BFGS",
                      start = c(alpha = unname(coef(m_start)[1] + 1) ,
                                beta = unname(coef(m_start)[2]), 
                                phi = 1/sigma(m_start)))

summary(wynik_ztnegbin)
```

```{r}
AIC(wynik_poisson)
```

```{r}
list(wynik_lichun,wynik_poisson, wynik_tpoisson, wynik_t2poisson, wynik_negbin, wynik_ztnegbin) %>%
  set_names(c("lichun", "poisson", "tpoisson", "t2poisson", "negbin", "ztnegbin")) %>%
  map_df( ~ data.frame(N_lower = sum(dt2019_1$N^coefci(.x)[1,1]),
                       N = sum(dt2019_1$N^coef(.x)[1]),
                       N_upper = sum(dt2019_1$N^coefci(.x)[1,2]),
                       AIC = AIC(.x)),
          .id = "model")
```


Separate models for sexes

```{r}
dt2019_1_females <- m %>% filter(half == "I", sex == 2) %>%
  rename(m = border,
         N = pesel,
         n = police_all)

wynik <- maxLik(logLik = ll, 
                grad = grad, 
                hess = hess,
                data = dt2019_1_females, 
                method = "NR",
                start = c(alpha = coef(m0)[1]+1, beta = coef(m0)[2], phi = 1/sigma(m0)))
summary(wynik)
AIC(wynik)
confint(wynik, level = 0.95)

sum(dt2019_1$N^0.5992235)
sum(dt2019_1$N^coef(wynik)[1]) ## 5525.491 + 5142.178
sum(dt2019_1$N^0.7317360)
```

Bootstrap

```{r}
eta_b <- (dt2019_1$n/dt2019_1$N)^0.58974*dt2019_1$N^0.66560*rgamma(n = nrow(dt2019_1), shape = 1, rate = 1.59143)
m_b <- rnbinom(n = nrow(dt2019_1), mu = eta, size = 1)+1

df <- data.frame(m = m_b, n = dt2019_1$n, N = dt2019_1$N)

wynik_b <- maxLik(logLik = ll, 
                  grad = grad, 
                  hess = hess,
                  data = df, 
                  method = "NR",
                  start = c(alpha = coef(m0)[1]+1, beta = coef(m0)[2], phi = 1/sigma(m0)))

summary(wynik_b)

```
