==> notebooks/3-final-models.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

Load the packages

```{r}
library(tidyverse)
library(maxLik)
library(actuar)
library(mvnfast)
library(SPIn)
library(patchwork)
```

Read the data

```{r}
data_for_model <- readRDS(file = "../data/data-for-model-aggregated.rds") %>%
  rename(m = border,
         N = pesel,
         n = police_all) %>%
  mutate(ukr = as.numeric(country_new == "Ukraine"),
         sex = as.numeric(sex == "Males"))

data_for_model
```

For the first half 

```{r}
data <- subset(data_for_model, half == "First half of 2019")

m0 <- estimate_pop(alpha = ~ 1, beta = ~ 1 , data = data, family = "NB2")
m1 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , data = data, family = "NB2")
m2 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ ukr , data = data, family = "NB2")
m3 <- estimate_pop(alpha = ~ 1 + sex, beta = ~ 1 , data = data, family = "NB2")
m4 <- estimate_pop(alpha = ~ 1 , beta = ~ sex , data = data, family = "NB2")
m5 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr , data = data, family = "NB2")
m6 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr + sex , data = data, family = "NB2")
m7 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 , data = data, family = "NB2")

nb2_loglik <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "maximum")
nb2_aic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "AIC")
nb2_bic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "BIC")
nb2_xi <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "xi")


m0 <- estimate_pop(alpha = ~ 1, beta = ~ 1 , data = data, family = "t0-NB2")
m1 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , data = data, family = "t0-NB2")
m2 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ ukr , data = data, family = "t0-NB2")
m3 <- estimate_pop(alpha = ~ 1 + sex, beta = ~ 1 , data = data, family = "t0-NB2")
m4 <- estimate_pop(alpha = ~ 1 , beta = ~ sex , data = data, family = "t0-NB2")
m5 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr , data = data, family = "t0-NB2")
m6 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr + sex , data = data, family = "t0-NB2")
m7 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 , data = data, family = "t0-NB2")

t0nb2_loglik <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "maximum")
t0nb2_aic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "AIC")
t0nb2_bic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "BIC")
t0nb2_xi <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "xi")

## PO

m0 <- estimate_pop(alpha = ~ 1, beta = ~ 1 , data = data, family = "PO")
m1 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , data = data, family = "PO")
m2 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ ukr , data = data, family = "PO")
m3 <- estimate_pop(alpha = ~ 1 + sex, beta = ~ 1 , data = data, family = "PO")
m4 <- estimate_pop(alpha = ~ 1 , beta = ~ sex , data = data, family = "PO")
m5 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr , data = data, family = "PO")
m6 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr + sex , data = data, family = "PO")
m7 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 , data = data, family = "PO")

po_loglik <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "maximum")
po_aic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "AIC")
po_bic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "BIC")
po_xi <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "xi")

m0 <- estimate_pop(alpha = ~ 1, beta = ~ 1 , data = data, family = "t0-PO")
m1 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , data = data, family = "t0-PO")
m2 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ ukr , data = data, family = "t0-PO")
m3 <- estimate_pop(alpha = ~ 1 + sex, beta = ~ 1 , data = data, family = "t0-PO")
m4 <- estimate_pop(alpha = ~ 1 , beta = ~ sex , data = data, family = "t0-PO")
m5 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr , data = data, family = "t0-PO")
m6 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr + sex , data = data, family = "t0-PO")
m7 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 , data = data, family = "t0-PO")

t0po_loglik <- sapply(list(m0,m1,m2,m3,m4,m5,m6, m7), "[[", "maximum")
t0po_aic <- sapply(list(m0,m1,m2,m3,m4,m5,m6, m7), "[[", "AIC")
t0po_bic <- sapply(list(m0,m1,m2,m3,m4,m5,m6, m7), "[[", "BIC")
t0po_xi <- sapply(list(m0,m1,m2,m3,m4,m5,m6, m7), "[[", "xi")

##Table with results

bind_rows(
  data.frame(model = "PO", x = 1:8, loglik=po_loglik, AIC=po_aic, BIC=po_bic, xi=po_xi),
  data.frame(model = "ztPO", x = 1:8, loglik=t0po_loglik, AIC=t0po_aic, BIC=t0po_bic, xi=t0po_xi),
  data.frame(model = "NB2", x = 1:8, loglik=nb2_loglik, AIC=nb2_aic, BIC=nb2_bic, xi=nb2_xi) ,
  data.frame(model = "ztNB2", x = 1:8, loglik=t0nb2_loglik, AIC=t0nb2_aic, BIC=t0nb2_bic, xi=t0nb2_xi)
) -> table_first_half
```

```{r}
data <- subset(data_for_model, half == "Second half of 2019")

m0 <- estimate_pop(alpha = ~ 1, beta = ~ 1 , data = data, family = "NB2")
m1 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , data = data, family = "NB2")
m2 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ ukr , data = data, family = "NB2")
m3 <- estimate_pop(alpha = ~ 1 + sex, beta = ~ 1 , data = data, family = "NB2")
m4 <- estimate_pop(alpha = ~ 1 , beta = ~ sex , data = data, family = "NB2")
m5 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr , data = data, family = "NB2")
m6 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr + sex , data = data, family = "NB2")
m7 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 , data = data, family = "NB2")

nb2_loglik <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "maximum")
nb2_aic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "AIC")
nb2_bic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "BIC")
nb2_xi <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "xi")


m0 <- estimate_pop(alpha = ~ 1, beta = ~ 1 , data = data, family = "t0-NB2")
m1 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , data = data, family = "t0-NB2")
m2 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ ukr , data = data, family = "t0-NB2")
m3 <- estimate_pop(alpha = ~ 1 + sex, beta = ~ 1 , data = data, family = "t0-NB2")
m4 <- estimate_pop(alpha = ~ 1 , beta = ~ sex , data = data, family = "t0-NB2")
m5 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr , data = data, family = "t0-NB2")
m6 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr + sex , data = data, family = "t0-NB2")
m7 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 , data = data, family = "t0-NB2")

t0nb2_loglik <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "maximum")
t0nb2_aic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "AIC")
t0nb2_bic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "BIC")
t0nb2_xi <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "xi")

## PO

m0 <- estimate_pop(alpha = ~ 1, beta = ~ 1 , data = data, family = "PO")
m1 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , data = data, family = "PO")
m2 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ ukr , data = data, family = "PO")
m3 <- estimate_pop(alpha = ~ 1 + sex, beta = ~ 1 , data = data, family = "PO")
m4 <- estimate_pop(alpha = ~ 1 , beta = ~ sex , data = data, family = "PO")
m5 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr , data = data, family = "PO")
m6 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr + sex , data = data, family = "PO")
m7 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 , data = data, family = "PO")

po_loglik <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "maximum")
po_aic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "AIC")
po_bic <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "BIC")
po_xi <- sapply(list(m0,m1,m2,m3,m4,m5,m6,m7), "[[", "xi")

m0 <- estimate_pop(alpha = ~ 1, beta = ~ 1 , data = data, family = "t0-PO")
m1 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , data = data, family = "t0-PO")
m2 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ ukr , data = data, family = "t0-PO")
m3 <- estimate_pop(alpha = ~ 1 + sex, beta = ~ 1 , data = data, family = "t0-PO")
m4 <- estimate_pop(alpha = ~ 1 , beta = ~ sex , data = data, family = "t0-PO")
m5 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr , data = data, family = "t0-PO")
m6 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 + ukr + sex , data = data, family = "t0-PO")
m7 <- estimate_pop(alpha = ~ 1 + ukr + sex , beta = ~ 1 , data = data, family = "t0-PO")

t0po_loglik <- sapply(list(m0,m1,m2,m3,m4,m5,m6, m7), "[[", "maximum")
t0po_aic <- sapply(list(m0,m1,m2,m3,m4,m5,m6, m7), "[[", "AIC")
t0po_bic <- sapply(list(m0,m1,m2,m3,m4,m5,m6, m7), "[[", "BIC")
t0po_xi <- sapply(list(m0,m1,m2,m3,m4,m5,m6, m7), "[[", "xi")

##Table with results

bind_rows(
  data.frame(model = "PO", x = 1:8, loglik=po_loglik, AIC=po_aic, BIC=po_bic, xi=po_xi),
  data.frame(model = "ztPO", x = 1:8, loglik=t0po_loglik, AIC=t0po_aic, BIC=t0po_bic, xi=t0po_xi),
  data.frame(model = "NB2", x = 1:8, loglik=nb2_loglik, AIC=nb2_aic, BIC=nb2_bic, xi=nb2_xi) ,
  data.frame(model = "ztNB2", x = 1:8, loglik=t0nb2_loglik, AIC=t0nb2_aic, BIC=t0nb2_bic, xi=t0nb2_xi)
) -> table_second_half
```

```{r}
bind_rows(
  table_first_half %>% filter(x %in% c(1,2,4, 8)) ,
  table_second_half %>% filter(x %in% c(1,2,4, 8)) 
) %>%
  xtable(digits = 1) %>%
  print.xtable(include.rownames = F, caption = "afasf", caption.placement = "top", format.args = list(big.mark = ","))

```

Which model is the best? 

```{r}
t2nb_model_null_1 <- estimate_pop(alpha = ~ 1, beta = ~ 1 , 
                                  data = subset(data_for_model, half == "First half of 2019"), 
                                  family = "t0-NB2")
t2nb_model_ukr_1 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , 
                                 data = subset(data_for_model, half == "First half of 2019"), 
                                 family = "t0-NB2")

t2nb_model_null_2 <- estimate_pop(alpha = ~ 1, beta = ~ 1 , 
                                  data = subset(data_for_model, half == "Second half of 2019"), 
                                  family = "t0-NB2")
t2nb_model_ukr_2 <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , 
                                 data = subset(data_for_model, half == "Second half of 2019"), 
                                 family = "t0-NB2")

```

```{r}
bind_rows(
data.frame(coef(summary(t2nb_model_null_1))) %>% 
  rownames_to_column(var = "par") %>%
  dplyr::select(par, est = Estimate, sd = `Std..error`) %>%
  mutate(model = "NB2-null", year = "2019Y1",
         est = round(est, 3),
         sd = round(sd, 3),
         print = paste0(est, " (",sd, ")")) %>%
  dplyr::select(year, model, par, print) %>%
  spread(par, print),
data.frame(coef(summary(t2nb_model_ukr_1))) %>% 
  rownames_to_column(var = "par") %>%
  dplyr::select(par, est = Estimate, sd = `Std..error`) %>%
  mutate(model = "NB2-ukr", year = "2019Y1",
         est = round(est, 3),
         sd = round(sd, 3),
         print = paste0(est, " (",sd, ")")) %>%
  dplyr::select(year, model, par, print) %>%
  spread(par, print),
data.frame(coef(summary(t2nb_model_null_2))) %>% 
  rownames_to_column(var = "par") %>%
  dplyr::select(par, est = Estimate, sd = `Std..error`) %>%
  mutate(model = "NB2-null", year = "2019Y2",
         est = round(est, 3),
         sd = round(sd, 3),
         print = paste0(est, " (",sd, ")")) %>%
  dplyr::select(year, model, par, print) %>%
  spread(par, print),
data.frame(coef(summary(t2nb_model_ukr_2))) %>% 
  rownames_to_column(var = "par") %>%
  dplyr::select(par, est = Estimate, sd = `Std..error`) %>%
  mutate(model = "NB2-ukr", year = "2019Y2",
         est = round(est, 3),
         sd = round(sd, 3),
         print = paste0(est, " (",sd, ")")) %>%
  dplyr::select(year, model, par, print) %>%
  spread(par, print)
) %>%
  dplyr::select(year, model, alpha0=`alpha_(Intercept)`, alpha1=`alpha_ukr`,
                beta = `beta_(Intercept)`, phi = phi) %>%
  xtable() %>%
  print.xtable(include.rownames = F)
```

```{r}
sum(((t2nb_model_null_1$fitted-t2nb_model_null_1$observed)^2))/1000
sum(((t2nb_model_ukr_1$fitted-t2nb_model_ukr_1$observed)^2))/1000
sum(((t2nb_model_null_2$fitted-t2nb_model_null_2$observed)^2))/1000
sum(((t2nb_model_ukr_2$fitted-t2nb_model_ukr_2$observed)^2))/1000
```

```{r}
plot(sqrt(data$m), sqrt(t2nb_model_ukr_sex$fitted))
abline(a=0,b=1,col='red')
```


Datailed results

```{r}
data_for_model %>%
  mutate(m_hat = ifelse(half == "First half of 2019", t2nb_model_ukr_1$fitted, t2nb_model_ukr_2$fitted),
         xi_i = ifelse(half == "First half of 2019", N^(0.6492643+0.09518408*ukr), N^(0.6725698+0.1036779*ukr))) %>%
  group_by(half) %>%
  summarise(total = sum(xi_i),
            Ukraine = sum(xi_i*ukr),
            work_age = sum(xi_i*(age == "Working age")),
            nonwork_age = total-work_age,
            males = sum(xi_i*sex),
            females = total - males) %>%
  xtable(digits = 0) %>%
  print.xtable(include.rownames = F, caption = "afasf", caption.placement = "top", format.args = list(big.mark = ","))
```


Bootstrap and MSE

```{r}
N_1 <- subset(data_for_model, half == "First half of 2019") %>% dplyr::pull(N)
ukr_1 <- subset(data_for_model, half == "First half of 2019") %>% dplyr::pull(ukr)
n_1 <- subset(data_for_model, half == "First half of 2019") %>% dplyr::pull(n)
m_1 <- subset(data_for_model, half == "First half of 2019") %>% dplyr::pull(m)
N_2 <- subset(data_for_model, half != "First half of 2019") %>% dplyr::pull(N)
ukr_2 <- subset(data_for_model, half != "First half of 2019") %>% dplyr::pull(ukr)
n_2 <- subset(data_for_model, half != "First half of 2019") %>% dplyr::pull(n)


set.seed(123)
boot_eta_1 <- rmvn(n=500, mu = coef(t2nb_model_ukr_1), sigma = vcov(t2nb_model_ukr_1))
boot_xi_1 <- apply(boot_eta_1, 1, FUN = function(x) sum(N_1^(x[1]+x[2]*ukr_1)))
boot_m_1 <- apply(boot_eta_1, 1, FUN = function(x) {
  mu_b <- N_1^(x[1]+x[2]*ukr_1)*(n_1/N_1)^x[3]
  xxx <- rztnbinom(n = NROW(N_1), size = x[4],  prob = x[4]/(x[4]+mu_b))
  xxx
  })

boot_eta_1 <- list() 

for (i in c(1:500)) {
  suppressWarnings(
    mod_b <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , data = data.frame(m = boot_m_1[,i], n = n_1, N=N_1, ukr = ukr_1), 
                          family = "t0-NB2")
  )
  boot_eta_1[[i]] <- coef(mod_b)
  cat(i, ": ", mod_b$message, "\n")
}

## problem with 108 iter 
boot_eta_1_hat <- do.call('rbind', boot_eta_1)
boot_xi_1_hat <- apply(boot_eta_1_hat, 1, FUN = function(x) sum(N_1^(x[1]+x[2]*ukr_1)))
mse_1 <- mean( (boot_xi_1_hat-boot_xi_1)^2 ) 

###### second half
set.seed(123)
boot_eta_2 <- rmvn(n=500, mu = coef(t2nb_model_ukr_2), sigma = vcov(t2nb_model_ukr_2))
boot_xi_2 <- apply(boot_eta_2, 1, FUN = function(x) sum(N_2^(x[1]+x[2]*ukr_2)))
boot_m_2 <- apply(boot_eta_2, 1, FUN = function(x) {
  mu_b <- N_2^(x[1]+x[2]*ukr_2)*(n_2/N_2)^x[3]
  xxx <- rztnbinom(n = NROW(N_2), size = x[4],  prob = x[4]/(x[4]+mu_b))
  xxx
  })

boot_eta_2 <- list() 

for (i in c(1:500)) {
  suppressWarnings(
    mod_b <- estimate_pop(alpha = ~ 1 + ukr, beta = ~ 1 , data = data.frame(m = boot_m_2[,i], n = n_2, N=N_2, ukr = ukr_2), 
                          family = "t0-NB2")
  )
  boot_eta_2[[i]] <- coef(mod_b)
  cat(i, ": ", mod_b$message, "\n")
}

## problem with 211, 308,492 iter 
boot_eta_2_hat <- do.call('rbind', boot_eta_2) 
boot_xi_2_hat <- apply(boot_eta_2_hat, 1, FUN = function(x) sum(N_2^(x[1]+x[2]*ukr_2)))
mse_2 <- mean( (boot_xi_2_hat[-c(211, 308,492)]-boot_xi_2[-c(211, 308,492)])^2 ) 


```


```{r}
data.frame(
  zhang = c(lower = sum(N_1^(.582307672+-0.002792677*ukr_1)), upper= sum(N_1^(0.7162209+0.1931608*ukr_1))),
  spin = spin(boot_xi_1, lower = 0),
  quantile = quantile(boot_xi_1, probs = c(0.025, 0.975))) %>%
  xtable(digits = 1) %>%
  print.xtable(include.rownames = F, caption = "afasf", caption.placement = "top", format.args = list(big.mark = ","))

data.frame(zhang = c(lower = sum(N_2^(0.5930046+-0.0134986*ukr_2)), upper= sum(N_2^(0.7521351+0.2208545*ukr_2))),
  spin = spin(boot_xi_2, lower = 0),
  quantile = quantile(boot_xi_2, probs = c(0.025, 0.975))) %>%
  xtable(digits = 1) %>%
  print.xtable(include.rownames = F, caption = "afasf", caption.placement = "top", format.args = list(big.mark = ","))

```


```{r}
tibble(period = "31.03.2019", xi = boot_xi_1_hat/1000) %>%
  bind_rows(tibble(period = "30.09.2019", xi = boot_xi_2_hat/1000)) %>%
  ggplot(data = ., aes(x = xi)) + 
  geom_histogram(fill = NA, color = "black", binwidth = 5) + 
  geom_rug(alpha = 0.5) + 
  facet_wrap(~period, nrow = 1) +
  labs(x = "Estimated irregular population size (in thousands)", y = "Count") -> p_hist

ggsave(plot = p_hist, filename = "../figs/fig-hist-mse.png", width = 6, height = 3)
```


Residuals for the models

1 quarter

```{r}
kappa1 <- t2nb_model_ukr_1$estimate[4]
mu1 <- as.numeric(t2nb_model_ukr_1$fitted)
m1 <- t2nb_model_ukr_1$observed
data1 <- subset(data_for_model, half == "At 31.03.2019")
data1$resid <- ( 3/kappa1*( (1+kappa1*m1)^(2/3) - (1+kappa1*mu1)^(2/3) ) + 3*(m1^(2/3)-mu1^(2/3)) ) / (2*(mu1+kappa1*mu1^2)^(1/6))
data1$mu <- mu1
data1$label <- ifelse(abs(data1$resid) > 5, paste(data1$country_new, data1$sex, data1$age, sep = ", "), "")
```

Plots

```{r}
ggplot(data = data1, aes(x = sqrt(border), sqrt(mu))) + 
  geom_point() + geom_abline(slope = 1, intercept = 0, linetype = "dashed")  + 
  labs(x = expression(sqrt(m)), y = expression(sqrt(hat(mu))), title = "Comparison of observed and fitted values") -> p1resid

ggplot(data = data1, aes(sample = resid)) + 
  stat_qq() + stat_qq_line() +  
  labs(x = "Theoretical quantile", y = "Sample quantile", title = "Normal Q-Q plot for Anscombe residuals") -> p2resid

ggplot(data = data1, aes(x = sqrt(mu), y = abs(resid))) + 
  geom_point() + geom_smooth(se = FALSE, linetype = "dashed", color = "black")  +
  labs(x = expression(sqrt(mu)), y = expression(abs(hat(r))), title = "Comparison of squared fitted and absolute \nAnscombe residuals") -> p3resid

ggplot(data = data1, aes(x = sqrt(mu), y = resid)) + 
  geom_point() + geom_smooth(se = FALSE, linetype = "dashed", color = "black")  +
  geom_hline(yintercept = 0, linetype = "dotted", color = "black") + 
  labs(x = expression(sqrt(mu)), y = expression(hat(r)), title = "Comparison of squared fitted and Anscombe \nresiduals") -> p4resid

p1resid + p2resid + p3resid + p4resid +
  plot_annotation(tag_levels = 'A') -> plot_all

ggsave(plot = plot_all, filename = "../figs/fig-resid-1-quart.png", width = 9, height = 9)
```


````{r}
kappa2 <- t2nb_model_ukr_2$estimate[4]
mu2 <- as.numeric(t2nb_model_ukr_2$fitted)
m2 <- t2nb_model_ukr_2$observed
data2 <- subset(data_for_model, half == "At 30.09.2019")
data2$resid <- ( 3/kappa2*( (1+kappa2*m2)^(2/3) - (1+kappa2*mu2)^(2/3) ) + 3*(m2^(2/3)-mu2^(2/3)) ) / (2*(mu2+kappa1*mu2^2)^(1/6))
data2$mu <- mu2

```


```{r}
ggplot(data = data2, aes(x = sqrt(border), sqrt(mu))) + 
  geom_point() + geom_abline(slope = 1, intercept = 0, linetype = "dashed")  + 
  labs(x = expression(sqrt(m)), y = expression(sqrt(hat(mu))), title = "Comparison of observed and fitted values") -> p1resid

ggplot(data = data2, aes(sample = resid)) + 
  stat_qq() + stat_qq_line() +  
  labs(x = "Theoretical quantile", y = "Sample quantile", title = "Normal Q-Q plot for Anscombe residuals") -> p2resid

ggplot(data = data2, aes(x = sqrt(mu), y = abs(resid))) + 
  geom_point() + geom_smooth(se = FALSE, linetype = "dashed", color = "black")  +
  labs(x = expression(sqrt(mu)), y = expression(abs(hat(r))), title = "Comparison of squared fitted and absolute \nAnscombe residuals") -> p3resid

ggplot(data = data2, aes(x = sqrt(mu), y = resid)) + 
  geom_point() + geom_smooth(se = FALSE, linetype = "dashed", color = "black")  +
  geom_hline(yintercept = 0, linetype = "dotted", color = "black") + 
  labs(x = expression(sqrt(mu)), y = expression(hat(r)), title = "Comparison of squared fitted and Anscombe \nresiduals") -> p4resid

p1resid + p2resid + p3resid + p4resid +
  plot_annotation(tag_levels = 'A') -> plot_all

ggsave(plot = plot_all, filename = "../figs/fig-resid-2-quart.png", width = 9, height = 9)
```


```{r}
data1 %>%
  arrange(-abs(resid))

data2 %>%
  arrange(-abs(resid))

```
